Change to code_symbols_tool
