+++ crates/assistant_tools/Cargo.toml
@@ -27,6 +27,7 @@ language.workspace = true
 language_model.workspace = true
 log.workspace = true
 project.workspace = true
+regex.workspace = true
 release_channel.workspace = true
 schemars.workspace = true
 serde.workspace = true
