+++ crates/assistant_tools/src/assistant_tools.rs
@@ -1,5 +1,6 @@
 mod bash_tool;
 mod batch_tool;
+mod code_symbols_tool;
 mod copy_path_tool;
 mod create_directory_tool;
 mod create_file_tool;
@@ -12,7 +13,6 @@ mod list_directory_tool;
 mod move_path_tool;
 mod now_tool;
 mod open_tool;
-mod outline_tool;
 mod path_search_tool;
 mod read_file_tool;
 mod regex_search_tool;
@@ -29,6 +29,7 @@ use move_path_tool::MovePathTool;
 
 use crate::bash_tool::BashTool;
 use crate::batch_tool::BatchTool;
+use crate::code_symbols_tool::CodeSymbolsTool;
 use crate::create_directory_tool::CreateDirectoryTool;
 use crate::create_file_tool::CreateFileTool;
 use crate::delete_path_tool::DeletePathTool;
@@ -39,7 +40,6 @@ use crate::find_replace_file_tool::FindReplaceFileTool;
 use crate::list_directory_tool::ListDirectoryTool;
 use crate::now_tool::NowTool;
 use crate::open_tool::OpenTool;
-use crate::outline_tool::OutlineTool;
 use crate::path_search_tool::PathSearchTool;
 use crate::read_file_tool::ReadFileTool;
 use crate::regex_search_tool::RegexSearchTool;
@@ -63,7 +63,7 @@ pub fn init(http_client: Arc<HttpClientWithUrl>, cx: &mut App) {
     registry.register_tool(ListDirectoryTool);
     registry.register_tool(NowTool);
     registry.register_tool(OpenTool);
-    registry.register_tool(OutlineTool);
+    registry.register_tool(CodeSymbolsTool);
     registry.register_tool(PathSearchTool);
     registry.register_tool(ReadFileTool);
     registry.register_tool(RegexSearchTool);
